pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "happler"
version = "0.0.1"
description = "A haplotype-based fine-mapping method"
authors = ["Arya Massarat"]
license = "MIT"
repository = "https://github.com/gymrek-lab/happler"
homepage = "https://github.com/gymrek-lab/happler"
documentation = "https://happler.readthedocs.io"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
click = ">=8.0.4"
networkx = ">=2.6.3"
scipy = ">=1.7.3"
pydot = ">=1.4.2"
pysam = ">=0.19.0"
haptools = ">=0.4.0"
statsmodels = ">=0.13.4"

[tool.poetry.group.docs.dependencies]
sphinx = ">=4.3.2"
sphinx-autodoc-typehints = ">=1.13.0"
sphinx-rtd-theme = ">=1.0.0"
numpydoc = ">=1.1.0"
sphinx-click = ">=3.0.2"

[tool.poetry.group.tests.dependencies]
pytest = ">=7.4.3"
black = ">=23.3.0"
ipython = ">=7.34.0"
coverage = {extras = ["toml"], version = ">=7.2.7"}
filprofiler = ">=2023.3.1"

[tool.poetry.scripts]
happler = 'happler.__main__:main'

[tool.black]
line-length = 90
preview = true
extend-exclude = "analysis/"

[tool.pytest.ini_options]
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
